REP-188: Move database metrics reporting into a dedicated class
Generating database metrics as part of the database healthcheck is a bit
strange - we have no control over how often healthchecks occur.

As a step towards making metrics collection a background asynchronous task,
move it to a separate class.
